refactor: Change File Dir
유저 관련 page, css, test 폴더 정리
- 추후 company 페이지 개발 고려
