+++ arc-crm/src/App.tsx
@@ -1,11 +1,11 @@
 import React, {JSX} from 'react';
 import { Routes, Route, Navigate } from 'react-router-dom';
-import LoginPage from './pages/LoginPage';
-import SignUpPage from './pages/SignUpPage';
+import LoginPage from './pages/user-pages/LoginPage';
+import SignUpPage from './pages/user-pages/SignUpPage';
 import CompanyPage from './pages/CompanyPage';
-import UserPage from './pages/UserPage';
-import MyInfoPage from './pages/MyInfoPage';
-import EditMyInfoPage from './pages/EditMyInfoPage';
+import UserPage from './pages/user-pages/UserPage';
+import MyInfoPage from './pages/user-pages/MyInfoPage';
+import EditMyInfoPage from './pages/user-pages/EditMyInfoPage';
 
 
 // 유효성 체크
