+++ arc-crm/src/pages/user-pages/EditMyInfoPage.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import '../styles/SignUpPage.css';
+import '../../styles/user-styles/SignUpPage.css';
 
 function EditMyInfoPage() {
     const navigate = useNavigate();
