+++ arc-crm/src/test/user-test/UserPage.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { MemoryRouter } from 'react-router-dom';
-import UserPage from '../pages/UserPage';
+import UserPage from '../../pages/user-pages/UserPage';
 
 // navigate 모킹
 const mockNavigate = jest.fn();
